small tweaks to hyperparameters and stat collection

+++ clear_knowledge.py
@@ -2,7 +2,6 @@
 
 import os
 import shutil
-import glob
 from stats import FILE_NAME as stats_file
 from test_environment import TEST_VIDEO_FILENAME as test_video_file
 
@@ -12,5 +11,8 @@ if os.path.exists(test_video_file):
 if os.path.exists(stats_file):
   os.remove(stats_file)
 
-shutil.rmtree('videos/')
-shutil.rmtree('checkpoints/')
+if os.path.exists('videos/'):
+  shutil.rmtree('videos/')
+
+if os.path.exists('checkpoints/'):
+  shutil.rmtree('checkpoints/')
